Add all current api calls to the docs
This document has been created from docs_md command. Its just to list whats currently available.
The docs are as good as plexpy own docs :(
I think it would be wise to switch to google style docs in the entire project as its much simpler to parse to something decent.