Log IP address in web auth logs
This commit adds the IP address of the attempted user login to all applicable logs within web auth.

I also created a helper class to abstract some of the logic of extracting pieces of the HTTP request headers. This should help cut down on any future duplication.

As well, I modified a couple of the log messages to read a bit better with this new output

This resolves issue #1111