Enhance reverse proxy support
* Move `tools.proxy.on` setting to from tree conf to options dict
because it was not present during redirects; use boolean value instead
of 0/1
* Create `redirect()` function to handle all redirects
* If `tools.proxy.on` is enabled in cherrypy config, check if
`X-Forwarded-Host` header is present (Apache). If not, fall back on
`X-Host` (lighttpd), `Origin` (Squid), or `Host` (nginx).

Note: I could not get a single `from plexpy.webserve import redirect`
statement to work at the top of `webauth.py`. I tried it all different
ways, and I could only get it to work by importing prior to each call to
`redirect()`. If you have any suggestions to make it work, I will be
happy to change it to a single import statement.