add option to parse MQTT message as JSON
same as webhooks. this enables proper escaping if one of the tags
(like a show's title) includes a " or \ character.

example message body:

`{ "title": "{title}", "media_type": "{media_type}" }`

and resulting message, for a TV episode with double quotes:

`{"title": "Ghosts - Ghosts (S2E1): The \"Grey\" Lady", "media_type": "episode"}`

(in this example, incl_subject is turned off, so that it's not nested inside
yet another JSON object, but if it's not, it does properly double-escape
the backslash.